+++ DarkStar.Client/Models/Tiles/Tile.cs
@@ -8,10 +8,20 @@ public class Tile
     public uint TileId { get; set; }
     public PointPosition Position { get; set; }
 
+    public TileOrientation Orientation { get; set; } = TileOrientation.West;
+
     public Tile(string id, uint tileId, PointPosition position)
     {
         Id = id;
         TileId = tileId;
         Position = position;
     }
+}
+
+public enum TileOrientation
+{
+    North,
+    South,
+    East,
+    West
 }