+++ DarkStar.Client/PageViewModels/RenderPageViewModel.cs
@@ -1,4 +1,7 @@
-﻿using System.Threading.Tasks;
+﻿using System.Reactive;
+using System.Threading.Tasks;
+using Avalonia;
+using Avalonia.Input;
 using DarkStar.Api.World.Types.Map;
 using DarkStar.Client.Attributes;
 using DarkStar.Client.Controls;
@@ -8,8 +11,12 @@ using DarkStar.Client.Services;
 using DarkStar.Client.ViewModels;
 using DarkStar.Network.Protocol.Interfaces.Messages;
 using DarkStar.Network.Protocol.Map;
+using DarkStar.Network.Protocol.Messages.Common;
+using DarkStar.Network.Protocol.Messages.Players;
 using DarkStar.Network.Protocol.Messages.Triggers.Npc;
 using DarkStar.Network.Protocol.Types;
+using FastEnumUtility;
+using ReactiveUI;
 
 namespace DarkStar.Client.PageViewModels;
 
@@ -17,23 +24,70 @@ namespace DarkStar.Client.PageViewModels;
 public class RenderPageViewModel : PageViewModelBase
 {
     private readonly GraphicEngineRender _graphicEngineRender;
-    private readonly ServiceContext _serviceContext;
 
+    private readonly ServiceContext _serviceContext;
+    public ReactiveCommand<string, Unit> MoveCharacterCommand { get; set; }
+
+    public ReactiveCommand<Unit, Task> PerformActionCommand { get; set; }
 
     public RenderPageViewModel(GraphicEngineRender graphicEngineRender, ServiceContext serviceContext)
     {
         _graphicEngineRender = graphicEngineRender;
         _serviceContext = serviceContext;
 
-        _serviceContext.NetworkClient.SubscribeToMessage<MapResponseMessage>(DarkStarMessageType.MapResponse, OnMapResponse);
-        _serviceContext.NetworkClient.SubscribeToMessage<NpcMovedResponseMessage>(DarkStarMessageType.NpcMovedResponse, OnNpcMoved);
+        serviceContext.NetworkClient.SubscribeToMessage<MapResponseMessage>(DarkStarMessageType.MapResponse, OnMapResponse);
+        serviceContext.NetworkClient.SubscribeToMessage<PlayerMoveResponseMessage>(
+            DarkStarMessageType.PlayerMoveResponse,
+            OnPlayerMoved
+        );
+        serviceContext.NetworkClient.SubscribeToMessage<PlayerDataResponseMessage>(
+            DarkStarMessageType.PlayerDataResponse,
+            OnPlayerData
+        );
+        serviceContext.NetworkClient.SubscribeToMessage<NpcMovedResponseMessage>(
+            DarkStarMessageType.NpcMovedResponse,
+            OnNpcMoved
+        );
+
+        MoveCharacterCommand = ReactiveCommand.Create<string, Unit>(
+            s =>
+            {
+                MoveCharacter(s);
+                return Unit.Default;
+            });
+
+        PerformActionCommand = ReactiveCommand.Create(PerformAction);
+
+    }
+
+    private Task PerformAction()
+    {
+       //return _serviceContext.NetworkClient.SendMessageAsync()
+       return Task.CompletedTask;
+    }
+
+    private Task OnPlayerMoved(IDarkStarNetworkMessage arg)
+    {
+        var message = (PlayerMoveResponseMessage)arg;
+
+        _graphicEngineRender.MoveTile(MapLayer.Players, message.PlayerId, message.Position);
+
+        return Task.CompletedTask;
+    }
+
+    private Task OnPlayerData(IDarkStarNetworkMessage arg)
+    {
+        var message = (PlayerDataResponseMessage)arg;
+        _graphicEngineRender.AddPlayer(message.PlayerId.ToString(), message.TileId, message.Position);
+
+        return Task.CompletedTask;
     }
 
     private Task OnNpcMoved(IDarkStarNetworkMessage arg)
     {
         var message = (NpcMovedResponseMessage)arg;
 
-        _graphicEngineRender.MoveTile(MapLayer.Creatures, message.NpcId.ToString(), message.Position);
+        _graphicEngineRender.MoveTile(MapLayer.Creatures, message.NpcId, message.Position);
 
         return Task.CompletedTask;
     }
@@ -57,6 +111,22 @@ public class RenderPageViewModel : PageViewModelBase
             _graphicEngineRender.AddTile(MapLayer.Creatures, new Tile(npc.Id.ToString(), npc.TileType, npc.Position));
         }
 
+        foreach (var gameObject in message.GameObjectsLayer)
+        {
+            _graphicEngineRender.AddTile(
+                MapLayer.Objects,
+                new Tile(gameObject.Id.ToString(), gameObject.TileType, gameObject.Position)
+            );
+        }
+
+        foreach (var item in message.ItemsLayer)
+        {
+            _graphicEngineRender.AddTile(
+                MapLayer.Items,
+                new Tile(item.Id.ToString(), item.TileType, item.Position)
+            );
+        }
+
         return Task.CompletedTask;
     }
 
@@ -64,4 +134,30 @@ public class RenderPageViewModel : PageViewModelBase
     {
         renderControl.RenderAction = _graphicEngineRender.RenderAction;
     }
+
+    public void DispatchKey(Key key, KeyModifiers modifiers)
+    {
+    }
+
+    public void DispatchMouseMove(Point point)
+    {
+    }
+
+    public void DispatchMouseWheel(Vector delta)
+    {
+    }
+
+    public void MoveCharacter(string direction)
+    {
+        var movement = FastEnum.Parse<MoveDirectionType>(direction);
+        _ = Task.Run(
+            () =>
+            {
+                _serviceContext.NetworkClient.SendMessageAsync(new PlayerMoveRequestMessage(movement));
+            }
+        );
+
+    }
+
+
 }