+++ DarkStar.Client/PageViews/RenderPageView.axaml.cs
@@ -15,18 +15,53 @@ public partial class RenderPageView : UserControl
 
     public RenderPageView()
     {
+
         InitializeComponent();
     }
 
     private void InitializeComponent()
     {
+
         AvaloniaXamlLoader.Load(this);
         _canvas = this.GetControl<RenderControl>("RenderControl");
+
+
         DataContextChanged += (sender, args) =>
         {
             if (DataContext is RenderPageViewModel renderPageViewModel)
             {
                 renderPageViewModel.SetupRenderControl(_canvas);
+            }
+        };
+
+        KeyDown += (sender, args) =>
+        {
+            if (DataContext is RenderPageViewModel renderPageViewModel)
+            {
+                renderPageViewModel.DispatchKey(args.Key, args.KeyModifiers);
+            }
+        };
+
+        PointerWheelChanged += (sender, args) =>
+        {
+            if (DataContext is RenderPageViewModel renderPageViewModel)
+            {
+                renderPageViewModel.DispatchMouseWheel(args.Delta);
+            }
+        };
+
+        PointerMoved += (sender, args) =>
+        {
+            if (DataContext is RenderPageViewModel renderPageViewModel)
+            {
+                renderPageViewModel.DispatchMouseMove(args.GetPosition(this));
+            }
+        };
+
+        PointerPressed += (sender, args) =>
+        {
+            if (DataContext is RenderPageViewModel renderPageViewModel)
+            {
             }
         };
     }